requirements.txt
flask
flask-cors
gTTS
pillow
opencv-python
pytesseract
numpy
tensorflow
easyocr
torch